refactor(Company) : divide Company as component
