+++ src/components/Company/styles/CompanyChart.scss
@@ -1,0 +1,25 @@
+.company-chart {
+	width: 45%;
+	height: 100%;
+	display: flex;
+	flex-direction: column;
+	margin-left: 10px;
+
+	@media screen and (max-width: 820px) {
+		margin-left: 0px;
+	}
+
+	.company-title {
+		background-color: white;
+		margin-bottom: 10px;
+
+		@media screen and (max-width: 820px) {
+			font-size: 2em;
+		}
+	}
+
+	@media screen and (max-width: 820px) {
+		width: 100%;
+		height: 400px;
+	}
+}
