+++ src/components/Company/styles/RecommendList.scss
@@ -1,0 +1,3 @@
+.recommend-list {
+	width: 25%;
+}
