--- src/pages/Company/css/Company.scss
@@ -1,60 +1,0 @@
-.container {
-	height: calc(100vh - 96px);
-	display: flex;
-	padding: 32px;
-	justify-content: center;
-	font-size: 1vw;
-
-	@media screen and (max-width: 820px) {
-		width: 100%;
-		height: 100%;
-		flex-direction: column;
-		padding: 24px 24px;
-	}
-
-	.company-chart-view {
-		width: 70%;
-		height: 100%;
-		display: flex;
-		flex-direction: row;
-
-		@media screen and (max-width: 820px) {
-			width: 100%;
-			flex-direction: column-reverse;
-		}
-
-		.company-chart {
-			width: 65%;
-			height: 100%;
-			display: flex;
-			flex-direction: column;
-			margin-left: 10px;
-
-			.company-title {
-				background-color: white;
-				margin-bottom: 10px;
-
-				@media screen and (max-width: 820px) {
-					font-size: 2em;
-				}
-			}
-
-			@media screen and (max-width: 820px) {
-				width: 100%;
-				height: 400px;
-				margin-left: 0px;
-			}
-		}
-	}
-
-	.vertical-line {
-		width: 1px;
-		background-color: #0c2a6b;
-		margin: 20px;
-		border-radius: 2px;
-
-		@media screen and (max-width: 820px) {
-			display: none;
-		}
-	}
-}
