+++ src/pages/Company/styles/Company.scss
@@ -1,0 +1,26 @@
+.container {
+	height: calc(100vh - 96px);
+	display: flex;
+	padding: 32px;
+	justify-content: center;
+	font-size: 1vw;
+
+	@media screen and (max-width: 820px) {
+		width: 100%;
+		height: 100%;
+		flex-direction: column;
+		padding: 24px 24px;
+	}
+
+	.company-chart-view {
+		width: 70%;
+		height: 100%;
+		display: flex;
+		flex-direction: row;
+
+		@media screen and (max-width: 820px) {
+			width: 100%;
+			flex-direction: column-reverse;
+		}
+	}
+}
